split server into classes
chopped the server code up to make it clearer taking advantage of classes.
