+++ package.json
@@ -5,8 +5,8 @@
   "description": "A Physical Space for Twitch Frogs to Hang and Chat.",
   "main": "server/Server.js",
   "scripts": {
-    "dev": "nodemon server/Server.js",
-    "run": "pm2 start server/Server.js && pm2 monit",
+    "dev": "nodemon server/Run.js",
+    "run": "pm2 start server/Run.js && pm2 monit",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [
